separa la logica de crypto_bloc
